pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic_backup_explorer"
version = "0.1.0"
description = "A semantic search and sync tool for backups."
requires-python = ">=3.10, <3.14"
dependencies = [
    "gradio",
    "langchain",
    "chromadb>=0.5.0",
    "sentence-transformers",
    "llm-client @ git+https://github.com/dgaida/llm_client.git",
    "python-dotenv",
    "tqdm",
    "pydantic-settings",
    "pydantic>=2.0",
    "typing-extensions>=4.10.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "pre-commit",
    "mypy",
    "types-tqdm",
    "bandit",
    "safety"
]

[tool.ruff]
line-length = 127
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B"]
ignore = ["E501", "B008", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["semantic_backup_explorer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=semantic_backup_explorer --cov-report=term-missing"

[tool.setuptools]
packages = [
    "semantic_backup_explorer",
    "semantic_backup_explorer.cli",
    "semantic_backup_explorer.cli.ui",
    "semantic_backup_explorer.chunking",
    "semantic_backup_explorer.compare",
    "semantic_backup_explorer.core",
    "semantic_backup_explorer.indexer",
    "semantic_backup_explorer.rag",
    "semantic_backup_explorer.sync",
    "semantic_backup_explorer.utils"
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "llm_client.*",
    "gradio.*",
    "tqdm.*",
    "dotenv.*",
    "pydantic_settings.*"
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["semantic_backup_explorer"]
omit = ["*/tests/*", "*/__pycache__/*", "*/cli/ui/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
